DONE Started trying to add rebase stuff that just ain't gonna work. Finished delete functionality. Additive editing now works, and you can't subtract from existing graphs. Toasts are now made from the Toast module.

+++ src/components/LogoutModal.vue
@@ -3,13 +3,15 @@
 		<div id="box">
 			<h1>Logout</h1>
 			<p>
-				Logging out will delete your repository and GitHub auth token on the server, however it will remain on GitHub.
+				You can log off of the website, however your data and local repo will still remain on the server in case you ever want to make modifications to them.
 			</p>
 			<p>
-				For security purposes, I have elected not to delete the repository with the API, so you will be taken to the GitHub page where you can delete it if you'd like. 
+				You can also delete your data (GH access token) as well as your local repo. Your contribution history will still remain on GitHub so long as you don't delete your contribeautiful_data repository, however, if you want to modify your history again, you'll have to delete that repo from GitHub before re-authorizing.
 			</p>
+			<p>For security reasons, the server will not clone your GH repo once it has been created. If you delete your data, you will be taken to a page where you can choose to delete your GH repo.</p>
 			<p>
-				Deleting the repo from GitHub will remove your drawings from your contribution graph.
+				<input id="chkDelete" type="checkbox" v-model="delData" />
+				<label for="chkDelete">Delete your data?</label>
 			</p>
 			<button class="btn-red" @click="logout">Logout</button>
 			<button @click="cancel">Cancel</button>
@@ -21,12 +23,15 @@
 export default {
 	name: 'LogoutModal',
 	props: ['username'],
+	data() { return {delData: false};},
 	methods: {
 		async logout() {
-			await fetch(`${process.env.VUE_APP_SERVER_BASE_URL}/user/${localStorage.getItem('userID')}`, {method: 'DELETE'});
+			if(this.delData) {
+				await fetch(`${process.env.VUE_APP_SERVER_BASE_URL}/user/${localStorage.getItem('userID')}`, {method: 'DELETE'});
+				window.open(`https://github.com/${this.username}/contribeautiful_data/settings#danger-zone`, '_blank');
+			}
 			localStorage.removeItem('userID');
 			location.reload();
-			window.open(`https://github.com/${this.username}/contribeautiful_data/settings#danger-zone`, '_blank');
 		},
 		cancel() {
 			this.$el.parentNode.parentNode.removeChild(this.$el.parentNode);
